analysis/requirements.txt
asyncpg~=0.30.0
certifi~=2025.7.14
fastapi~=0.116.1
python-dotenv~=1.1.1
starlette~=0.47.2
pydantic~=2.11.7
ttkbootstrap~=1.15.0
pandas~=2.3.1
numpy~=2.3.1
scikit-learn~=1.7.1
aiohttp~=3.12.15
tqdm~=4.67.1